Alias creation bug fix

+++ core/serializers.py
@@ -12,8 +12,7 @@ from dateutil.relativedelta import *
 from django.contrib.auth.models import Permission, User
 from django.db.models import Q
 from django.utils import timezone
-from django.utils.timezone import make_aware
-from rest_framework import pagination, serializers
+from rest_framework import serializers
 from rest_framework.utils import model_meta
 
 from core import query
@@ -22,7 +21,7 @@ from core.models import (
     Operation, OperationStep, OperationDataColumnAlias, Review, ScheduledEvent,
     ScheduledEventRunInstance, Sector, Source, SourceColumnMap, Tag,
     Theme, UpdateHistory, FrozenData, SavedQueryData)
-
+from core.errors import AliasCreationError, AliasUpdateError
 
 class DataSerializer(serializers.BaseSerializer):
     """
@@ -170,6 +169,7 @@ class OperationSerializer(serializers.ModelSerializer):
             'created_on',
             'updated_on',
             'aliases',
+            'alias_creation_status'
         )
 
     def create(self, validated_data):
@@ -185,9 +185,11 @@ class OperationSerializer(serializers.ModelSerializer):
         operation.operation_query = query.build_query(operation=operation)
         operation.count_rows = True
         operation.save()
-        self.create_operation_data_aliases(operation)
-
-        return operation
+        try:
+            self.create_operation_data_aliases(operation)
+            return operation
+        except AliasCreationError:
+            raise AliasCreationError({'error_code': operation.alias_creation_status})
 
     def update(self, instance, validated_data):
         info = model_meta.get_field_info(instance)
@@ -223,41 +225,71 @@ class OperationSerializer(serializers.ModelSerializer):
         instance.operation_query = query.build_query(operation=instance)
         instance.count_rows = True
         instance.save()
-        self.update_operation_data_aliases(instance)
-
-        return instance
+        try:
+            self.update_operation_data_aliases(instance)
+            return instance
+        except AliasUpdateError:
+            raise AliasUpdateError({'error_code': instance.alias_creation_status})
 
     def create_operation_data_aliases(self, operation):
         count, data = query.query_table(operation, 1, 0, estimate_count=True)
-        try:
-            data_column_keys = data[0].keys()
-            first_step = operation.get_operation_steps()[0]
-            columns = SourceColumnMap.objects.filter(source=first_step.source, name__in=data_column_keys)
-            for column in data_column_keys:
-                matching = columns.filter(name=column).first()
-                alias = self.create_operation_alias(operation, column, matching.alias if matching else column)
-                alias.save()
-        except: # FIXME: handle specific errors
-            pass
-
-    def update_operation_data_aliases(self, operation):
-        count, data = query.query_table(operation, 1, 0, estimate_count=True)
-        try:
-            data_column_keys = data[0].keys()
-            first_step = operation.get_operation_steps()[0]
-            columns = SourceColumnMap.objects.filter(source=first_step.source, name__in=data_column_keys)
-            # delete obsolete aliases
-            OperationDataColumnAlias.objects.filter(operation=operation).exclude(column_name__in=data_column_keys).delete()
-            for column in data_column_keys:
-                existing_alias = OperationDataColumnAlias.objects.filter(operation=operation, column_name=column).first()
-                if not existing_alias:
+        operation.alias_creation_status = 'p'
+        operation.save()
+        if data:
+            try:
+                data_column_keys = data.keys() if isinstance(data, dict) else data[0].keys()
+                first_step = operation.get_operation_steps()[0]
+                columns = SourceColumnMap.objects.filter(source=first_step.source, name__in=data_column_keys)
+                for column in data_column_keys:
                     matching = columns.filter(name=column).first()
                     alias = self.create_operation_alias(operation, column, matching.alias if matching else column)
                     alias.save()
-        except: # FIXME: handle specific errors
-            pass
+                    operation.alias_creation_status = 'd'
+                    operation.save()
+            except:
+                operation.alias_creation_status = 'e'
+                operation.save()
+                raise AliasCreationError()
+        else:
+            operation.alias_creation_status = 'd'
+            operation.save()
+
+    def update_operation_data_aliases(self, operation):
+        count, data = query.query_table(operation, 1, 0, estimate_count=True)
+        operation.alias_creation_status = 'p'
+        operation.save()
+        if data:
+            try:
+                data_column_keys = data.keys() if isinstance(data, dict) else data[0].keys()
+                first_step = operation.get_operation_steps()[0]
+                columns = SourceColumnMap.objects.filter(source=first_step.source, name__in=data_column_keys)
+                # delete obsolete aliases
+                OperationDataColumnAlias.objects.filter(operation=operation).exclude(column_name__in=data_column_keys).delete()
+                for column in data_column_keys:
+                    existing_alias = OperationDataColumnAlias.objects.filter(operation=operation, column_name=column).first()
+                    if not existing_alias:
+                        matching = columns.filter(name=column).first()
+                        alias = self.create_operation_alias(operation, column, matching.alias if matching else column)
+                        alias.save()
+                        operation.alias_creation_status = 'd'
+                        operation.save()
+            except:
+                operation.alias_creation_status = 'e'
+                operation.save()
+                raise AliasUpdateError()
+        else:
+            operation.alias_creation_status = 'd'
+            operation.save()
 
     def create_operation_alias(self, operation, column_name, column_alias):
+        if not column_alias:
+            name_array = column_name.split('_')
+            capitalized_name_array = []
+            for word in name_array:
+                capitalized_word = word.capitalize()
+                capitalized_name_array.append(capitalized_word)
+            column_alias = ' '.join(capitalized_name_array)
+
         return OperationDataColumnAlias.objects.create(
             operation=operation, column_name=column_name, column_alias=column_alias)
